perf: Allow AUX sampling for multiple events
Right now, only one perf event in a context at a time can be annotated with
AUX data, because AUX pmus will normally have only one event scheduled at a
time, however it should be possible to annotate several events with similar
configuration (wrt exclude_{hv,idle,user,kernel}, config and other event
attribute fields) using the same AUX event.

This patch changes the behavior so that every time, before a kernel counter
is created for AUX sampling, we first look for an existing counter that is
suitable in terms of configuration and we use it to annotate the new event
as well. If no such event is found, we fall back to the old path and try
to allocate a new kernel counter.

Signed-off-by: Alexander Shishkin <alexander.shishkin@linux.intel.com>