perf tools: Only override the default :tid comm entry
Events may still be ordered even if there are no
timestamps e.g. if the data is recorded per-thread.

Also synthesized COMM events have a timestamp of zero.

Consequently it is better to keep comm entries even
if they have a timestamp of zero.

However, when a struct thread is created the command
string is not known and a comm entry with a string
of the form ":<tid>" is used.  In that case
thread->comm_set is false and the comm entry should
be overridden.

Signed-off-by: Adrian Hunter <adrian.hunter@intel.com>